+++ model_fetch/lib/src/model_by_filter/model_by_filter_bloc.dart
@@ -47,6 +47,8 @@ abstract class ModelByFilterBloc<
   Future<void> dispose() async {
     await _statesController.close();
   }
+
+  void reload();
 }
 
 class ModelByFilterState<I, T extends WithId<I>> {